# artifacts/api-server/src/routes/suppliers.ts
import { Router, type IRouter } from "express";
import { db, suppliersTable } from "@workspace/db";
import { eq } from "drizzle-orm";
import { requireAuth } from "../middlewares/requireAuth";
import { CreateSupplierBody, UpdateSupplierBody, UpdateSupplierParams } from "@workspace/api-zod";

const router: IRouter = Router();

router.get("/suppliers", requireAuth, async (_req, res): Promise<void> => {
  const suppliers = await db
    .select()
    .from(suppliersTable)
    .orderBy(suppliersTable.name);
  res.json(suppliers);
});

router.post("/suppliers", requireAuth, async (req, res): Promise<void> => {
  const parsed = CreateSupplierBody.safeParse(req.body);
  if (!parsed.success) {
    res.status(400).json({ error: parsed.error.message });
    return;
  }

  const [supplier] = await db
    .insert(suppliersTable)
    .values(parsed.data)
    .returning();

  res.status(201).json(supplier);
});

router.patch("/suppliers/:id", requireAuth, async (req, res): Promise<void> => {
  const params = UpdateSupplierParams.safeParse(req.params);
  if (!params.success) {
    res.status(400).json({ error: params.error.message });
    return;
  }

  const parsed = UpdateSupplierBody.safeParse(req.body);
  if (!parsed.success) {
    res.status(400).json({ error: parsed.error.message });
    return;
  }

  const [supplier] = await db
    .update(suppliersTable)
    .set(parsed.data)
    .where(eq(suppliersTable.id, params.data.id))
    .returning();

  if (!supplier) {
    res.status(404).json({ error: "Supplier not found" });
    return;
  }

  res.json(supplier);
});

export default router;