# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * MotoTrack ERP/POS API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface OkResponse {
  ok: boolean;
}

export interface ErrorResponse {
  error: string;
}

export interface LoginInput {
  /** @minLength 1 */
  username: string;
  /** @minLength 1 */
  password: string;
}

export interface SetCompanyInput {
  companyId: number;
}

export interface AuthUser {
  id: number;
  username: string;
  role: string;
  /** @nullable */
  companyId?: number | null;
  /** @nullable */
  companyName?: string | null;
}

export interface Company {
  id: number;
  code: string;
  name: string;
  active: boolean;
}

export interface User {
  id: number;
  username: string;
  role: string;
  active: boolean;
  createdAt?: string;
}

export interface UserInput {
  /** @minLength 1 */
  username: string;
  /** @minLength 6 */
  password: string;
  role: string;
}

export interface UserUpdate {
  role?: string;
  active?: boolean;
  password?: string;
}

export interface Location {
  id: number;
  companyId: number;
  name: string;
  code: string;
  active?: boolean;
}

export interface LocationInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  code: string;
}

export interface StockGroup {
  id: number;
  companyId: number;
  name: string;
  code: string;
}

export interface StockItem {
  id: number;
  companyId: number;
  /** @nullable */
  stockGroupId?: number | null;
  name: string;
  code: string;
  uom: string;
  /** @nullable */
  sellingPrice?: number | null;
  /** @nullable */
  openingQty?: number | null;
  /** @nullable */
  openingRate?: number | null;
  /** @nullable */
  reorderLevel?: number | null;
  active?: boolean;
}

export interface StockItemInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  code: string;
  uom: string;
  stockGroupId?: number;
  sellingPrice?: number;
  openingQty?: number;
  openingRate?: number;
  reorderLevel?: number;
}

export interface StockItemUpdate {
  name?: string;
  code?: string;
  uom?: string;
  sellingPrice?: number;
  reorderLevel?: number;
  active?: boolean;
}

export interface Supplier {
  id: number;
  name: string;
  /** @nullable */
  contactPerson?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  address?: string | null;
  active?: boolean;
}

export interface SupplierInput {
  /** @minLength 1 */
  name: string;
  contactPerson?: string;
  phone?: string;
  email?: string;
  address?: string;
}

export interface SupplierUpdate {
  name?: string;
  contactPerson?: string;
  phone?: string;
  email?: string;
  address?: string;
  active?: boolean;
}

export interface Container {
  id: number;
  companyId: number;
  /** @nullable */
  supplierId?: number | null;
  /** @nullable */
  supplierName?: string | null;
  containerNumber: string;
  status: string;
  /** @nullable */
  importDate?: string | null;
  /** @nullable */
  totalCost?: number | null;
  createdAt?: string;
}

export interface ContainerInput {
  /** @minLength 1 */
  containerNumber: string;
  supplierId?: number;
  importDate?: string;
  freightCost?: number;
  fumigationCost?: number;
  otherCharges?: number;
}

export interface ContainerItem {
  id: number;
  containerId: number;
  /** @nullable */
  stockItemId?: number | null;
  itemName: string;
  quantity: number;
  ratePerUnit: number;
}

export interface ContainerDetail {
  id: number;
  companyId: number;
  /** @nullable */
  supplierId?: number | null;
  /** @nullable */
  supplierName?: string | null;
  containerNumber: string;
  status: string;
  /** @nullable */
  importDate?: string | null;
  /** @nullable */
  freightCost?: number | null;
  /** @nullable */
  fumigationCost?: number | null;
  /** @nullable */
  otherCharges?: number | null;
  /** @nullable */
  totalCost?: number | null;
  createdAt?: string;
  items?: ContainerItem[];
}

export interface Account {
  id: number;
  companyId: number;
  name: string;
  /** @nullable */
  code?: string | null;
  accountType: string;
  /** @nullable */
  balance?: number | null;
  active?: boolean;
}

export interface AccountInput {
  /** @minLength 1 */
  name: string;
  code?: string;
  accountType: string;
}

export interface Voucher {
  id: number;
  companyId: number;
  voucherNumber: string;
  voucherType: string;
  voucherDate: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  totalAmount?: number | null;
  createdAt?: string;
}

export interface VoucherEntryInput {
  accountId: number;
  debitAmount?: number;
  creditAmount?: number;
  narration?: string;
}

export interface VoucherInput {
  voucherType: string;
  voucherDate: string;
  description?: string;
  entries?: VoucherEntryInput[];
}

export interface VoucherEntry {
  id: number;
  voucherId: number;
  accountId: number;
  /** @nullable */
  accountName?: string | null;
  /** @nullable */
  debitAmount?: number | null;
  /** @nullable */
  creditAmount?: number | null;
  /** @nullable */
  narration?: string | null;
}

export interface VoucherDetail {
  id: number;
  companyId: number;
  voucherNumber: string;
  voucherType: string;
  voucherDate: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  totalAmount?: number | null;
  createdAt?: string;
  entries?: VoucherEntry[];
}

export interface Employee {
  id: number;
  companyId: number;
  /** @nullable */
  code?: string | null;
  firstName: string;
  lastName: string;
  employeeType?: string;
  /** @nullable */
  monthlySalary?: number | null;
  /** @nullable */
  phone?: string | null;
  active?: boolean;
  /** @nullable */
  hireDate?: string | null;
}

export interface EmployeeInput {
  /** @minLength 1 */
  firstName: string;
  /** @minLength 1 */
  lastName: string;
  code?: string;
  employeeType?: string;
  monthlySalary?: number;
  phone?: string;
  hireDate?: string;
}

export interface EmployeeUpdate {
  firstName?: string;
  lastName?: string;
  employeeType?: string;
  monthlySalary?: number;
  phone?: string;
  active?: boolean;
}

export interface DashboardSummary {
  totalStockValue: number;
  totalReceivables: number;
  totalPayables: number;
  totalCash: number;
  voucherCount: number;
  employeeCount: number;
  stockItemCount: number;
}

export interface StockAlert {
  id: number;
  name: string;
  code: string;
  currentQty: number;
  reorderLevel: number;
}

export type ListVouchersParams = {
date?: string;
type?: string;
};