# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: MotoTrack ERP/POS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with username and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current session user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/set-company:
    post:
      operationId: setCompany
      tags: [auth]
      summary: Set the active company for the session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetCompanyInput"
      responses:
        "200":
          description: Company set
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
...
  /stock-items:
    get:
      operationId: listStockItems
      tags: [stockItems]
      summary: List stock items for current company
      responses:
        "200":
          description: Stock items list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StockItem"
    post:
      operationId: createStockItem
      tags: [stockItems]
      summary: Create a stock item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StockItemInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StockItem"
  /stock-items/{id}:
    patch:
      operationId: updateStockItem
      tags: [stockItems]
      summary: Update a stock item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StockItemUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StockItem"
    delete:
      operationId: deleteStockItem
      tags: [stockItems]
      summary: Delete a stock item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /suppliers:
    get:
      operationId: listSuppliers
      tags: [suppliers]
      summary: List suppliers
      responses:
        "200":
          description: Suppliers list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Supplier"
    post:
      operationId: createSupplier
      tags: [suppliers]
      summary: Create a supplier
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SupplierInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
  /suppliers/{id}:
    patch:
      operationId: updateSupplier
      tags: [suppliers]
      summary: Update a supplier
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SupplierUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
  /containers:
    get:
      operationId: listContainers
      tags: [containers]
      summary: List containers for current company
      responses:
        "200":
          description: Containers list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Container"
    post:
      operationId: createContainer
      tags: [containers]
      summary: Create a container
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContainerInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Container"
  /containers/{id}:
    get:
      operationId: getContainer
      tags: [containers]
      summary: Get container details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Container details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContainerDetail"
...
  /employees:
    get:
      operationId: listEmployees
      tags: [employees]
      summary: List employees for current company
      responses:
        "200":
          description: Employees list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Employee"
    post:
      operationId: createEmployee
      tags: [employees]
      summary: Create an employee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
  /employees/{id}:
    patch:
      operationId: updateEmployee
      tags: [employees]
      summary: Update an employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
...
components:
  schemas:
...
    OkResponse:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # Auth
    LoginInput:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          minLength: 1
        password:
          type: string
          minLength: 1
    SetCompanyInput:
      type: object
      required: [companyId]
      properties:
        companyId:
          type: integer
    AuthUser:
      type: object
      required: [id, username, role]
      properties:
        id:
          type: integer
        username:
          type: string
        role:
          type: string
        companyId:
          type: ["integer", "null"]
        companyName:
          type: ["string", "null"]

    # Company
...
    StockItem:
      type: object
      required: [id, companyId, name, code, uom]
      properties:
        id:
          type: integer
        companyId:
          type: integer
        stockGroupId:
          type: ["integer", "null"]
        name:
          type: string
        code:
          type: string
        uom:
          type: string
        sellingPrice:
          type: ["number", "null"]
        openingQty:
          type: ["number", "null"]
        openingRate:
          type: ["number", "null"]
        reorderLevel:
          type: ["number", "null"]
        active:
          type: boolean
    StockItemInput:
      type: object
      required: [name, code, uom]
      properties:
        name:
          type: string
          minLength: 1
        code:
          type: string
          minLength: 1
        uom:
          type: string
        stockGroupId:
          type: integer
        sellingPrice:
          type: number
        openingQty:
          type: number
        openingRate:
          type: number
        reorderLevel:
          type: number
    StockItemUpdate:
      type: object
      properties:
        name:
          type: string
        code:
          type: string
        uom:
          type: string
        sellingPrice:
          type: number
        reorderLevel:
          type: number
        active:
          type: boolean

    # Supplier
    Supplier:
      type: object
      required: [id, name]
      properties:
        id:
          type: integer
        name:
          type: string
        contactPerson:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        active:
          type: boolean
    SupplierInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        contactPerson:
          type: string
        phone:
          type: string
        email:
          type: string
        address:
          type: string
    SupplierUpdate:
      type: object
      properties:
        name:
          type: string
        contactPerson:
          type: string
        phone:
          type: string
        email:
          type: string
        address:
          type: string
        active:
          type: boolean

    # Container
    Container:
      type: object
      required: [id, companyId, containerNumber, status]
      properties:
        id:
          type: integer
        companyId:
          type: integer
        supplierId:
          type: ["integer", "null"]
        supplierName:
          type: ["string", "null"]
        containerNumber:
          type: string
        status:
          type: string
        importDate:
          type: ["string", "null"]
        totalCost:
          type: ["number", "null"]
        createdAt:
          type: string
    ContainerInput:
      type: object
      required: [containerNumber]
      properties:
        containerNumber:
          type: string
          minLength: 1
        supplierId:
          type: integer
        importDate:
          type: string
        freightCost:
          type: number
        fumigationCost:
          type: number
        otherCharges:
          type: number
    ContainerDetail:
      type: object
      required: [id, companyId, containerNumber, status]
      properties:
        id:
          type: integer
        companyId:
          type: integer
        supplierId:
          type: ["integer", "null"]
        supplierName:
          type: ["string", "null"]
        containerNumber:
          type: string
        status:
          type: string
        importDate:
          type: ["string", "null"]
        freightCost:
          type: ["number", "null"]
        fumigationCost:
          type: ["number", "null"]
        otherCharges:
          type: ["number", "null"]
        totalCost:
          type: ["number", "null"]
        createdAt:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/ContainerItem"
    ContainerItem:
      type: object
      required: [id, containerId, itemName, quantity, ratePerUnit]
      properties:
        id:
          type: integer
        containerId:
          type: integer
        stockItemId:
          type: ["integer", "null"]
        itemName:
          type: string
        quantity:
          type: number
        ratePerUnit:
          type: number

    # Account
...
    Employee:
      type: object
      required: [id, companyId, firstName, lastName]
      properties:
        id:
          type: integer
        companyId:
          type: integer
        code:
          type: ["string", "null"]
        firstName:
          type: string
        lastName:
          type: string
        employeeType:
          type: string
        monthlySalary:
          type: ["number", "null"]
        phone:
          type: ["string", "null"]
        active:
          type: boolean
        hireDate:
          type: ["string", "null"]
    EmployeeInput:
      type: object
      required: [firstName, lastName]
      properties:
        firstName:
          type: string
          minLength: 1
        lastName:
          type: string
          minLength: 1
        code:
          type: string
        employeeType:
          type: string
        monthlySalary:
          type: number
        phone:
          type: string
        hireDate:
          type: string
    EmployeeUpdate:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        employeeType:
          type: string
        monthlySalary:
          type: number
        phone:
          type: string
        active:
          type: boolean

    # Dashboard
...