# lib/api-zod/src/generated/types/container.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * MotoTrack ERP/POS API
 * OpenAPI spec version: 0.1.0
 */

export interface Container {
  id: number;
  companyId: number;
  /** @nullable */
  supplierId?: number | null;
  /** @nullable */
  supplierName?: string | null;
  containerNumber: string;
  status: string;
  /** @nullable */
  importDate?: string | null;
  /** @nullable */
  totalCost?: number | null;
  createdAt?: string;
}