# lib/api-zod/src/generated/types/containerDetail.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * MotoTrack ERP/POS API
 * OpenAPI spec version: 0.1.0
 */
import type { ContainerItem } from './containerItem';

export interface ContainerDetail {
  id: number;
  companyId: number;
  /** @nullable */
  supplierId?: number | null;
  /** @nullable */
  supplierName?: string | null;
  containerNumber: string;
  status: string;
  /** @nullable */
  importDate?: string | null;
  /** @nullable */
  freightCost?: number | null;
  /** @nullable */
  fumigationCost?: number | null;
  /** @nullable */
  otherCharges?: number | null;
  /** @nullable */
  totalCost?: number | null;
  createdAt?: string;
  items?: ContainerItem[];
}